feat(cockpit/activity tree): add filtering by name and state
This commit adds a textfield on top of the activities tree who
can be used to quickly find an acivity instance.
If applicable, a set of checkboxes is also added in order to
refine the search by the state (running, completed or canceled).

This commit also improve the linting of some files and fix
a small styling issue (activity label ot taking the whole
available width of the tree leaf).

Related: CAM-1874, CAM-1895