fix(test-helper): change the way incidents are triggered
Before:
Jobs were triggered to throw an incident by setting the jobRetries to 0
this behavior disables the job for the jobexecutor.
Drawback: When the job executor is faster than the request there will be
incident messages for the incidents else the message will be null

New Behavior:
The job executor will execute the jobs and count down the jobRetries. The
test-helper is only polling if there are any jobs left which need to be
executed before the tests can be started

Related to: #CAM-5523